requirements.txt
aiogram==3.13.1
groq==0.11.0
aiohttp==3.10.10
python-dotenv==1.0.1
